% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRelations.R
\name{getAttribute}
\alias{getAttribute}
\title{Get attributes from a vocab object returned by icesVocab::getCodeDetail
compatibility icesVocab >= version 1.3.3}
\usage{
getAttribute(st.)
}
\arguments{
\item{st.}{an object returned by getCodeDetail}
}
\value{
A list with description and value for attributes
}
\description{
Get attributes from a vocab object returned by icesVocab::getCodeDetail
compatibility icesVocab >= version 1.3.3
}
\examples{
\dontrun{
tt <- icesVocab::getCodeDetail("Station", 1000)
getAttribute(st. = tt)
}
}
