% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRelations.R
\name{getListRelations}
\alias{getListRelations}
\title{Get a vector of values for each relation
compatibility version 1.2.0}
\usage{
getListRelations(st., relation_key.)
}
\arguments{
\item{st.}{an object returned by getCodeDetail}

\item{relation_key}{the parent key returned by getRelations()}
}
\value{
A vector of relation values
}
\description{
Get a vector of values for each relation
compatibility version 1.2.0
}
\examples{
\dontrun{
tt <- icesVocab::getCodeDetail("Station", 1000)
relation_key <- getRelations(tt)
getListRelations(st. = tt, relation_key. = "ISO_3166") # FI
getListRelations(st. = tt, relation_key. = "Station_DTYPE") # EW~EU~NU~CW
}
}
