% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getListStation.R
\name{getListStation}
\alias{getListStation}
\title{Get a table of stations.}
\usage{
getListStation(code = NULL, deprecated = NULL, date = NULL)
}
\arguments{
\item{code}{a vector of station codes}

\item{deprecated}{logical choosing TRUE or FALSE will only return the
deprecated or not deprecated records.}

\item{date}{A date or character like '2025-01-01' to choose only the station
modified after that date}
}
\value{
A data frame of stations.
}
\description{
Get a table of stations.
}
\examples{
\dontrun{
# get all stations entered in the last month
station <- getListStation(date=Sys.Date()-as.difftime(4,units="weeks"))
station <- getListStation(date=Sys.Date()-as.difftime(4,units="weeks"), deprecated = FALSE)
station <- getListStation(code = 1:10)
}
}
