% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRelations.R
\name{getRelations}
\alias{getRelations}
\title{Get all relations from a vocab object returned by icesVocab::getCodeDetail
compatibility version icesVocab > 1.3.3}
\usage{
getRelations(st.)
}
\arguments{
\item{st.}{an object returned by getCodeDetail}
}
\value{
A vector with unique key of relations (as returned by
the api in the parents element of the list)
}
\description{
Get all relations from a vocab object returned by icesVocab::getCodeDetail
compatibility version icesVocab > 1.3.3
}
\examples{
\dontrun{
tt <- icesVocab::getCodeDetail("Station", 1000)
getRelations(st. = tt)
# [1] "Station_DTYPE" "PURPM"         "Datasets"      "ISO_3166"
# [5] "WLTYP"         "PRGOV"         "EDMO"
}
}
