% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStationDetail.R
\name{getStationDetail}
\alias{getStationDetail}
\title{Get a station record.
note that field dataset exists in the API but is not present in the
exchange format}
\usage{
getStationDetail(.code)
}
\arguments{
\item{.code}{The code of the station}
}
\value{
A table with station values filled. In fields with
external relations are collated together for instance if a station
returns association with EW, EU and NU for Station_DTYPE these character
vectors will be collated together with a `~` returning EW~EU~NU
}
\description{
Get a station record.
note that field dataset exists in the API but is not present in the
exchange format
}
\examples{
\dontrun{
stdetail <- getStationDetail(.code = 1000)
}
}
